Revert "Pkc 69 v2"
