+++ src/components/CustomLabel/index.tsx
@@ -1,19 +1,13 @@
-import { Typography } from "@mui/material";
-import styles from "./label.module.scss";
+import { Typography } from '@mui/material';
 
 interface CustomLabelProps {
-  name: string;
-  required: boolean;
+    name: string;
+    required: boolean;
 }
 
 const CustomLabel = ({ name, required }: CustomLabelProps) => (
-  <Typography variant="body1" sx={{ py: 1 }} className={styles["title"]}>
-    {name}{" "}
-    {required && (
-      <span style={{ color: "red", position: "relative", bottom: "-0.2em" }}>
-        *
-      </span>
-    )}
+  <Typography variant="body1" sx={{ py: 1 }}>
+    {name} {required && <span style={{ color: 'red', position: 'relative', bottom: '-0.2em' }}>*</span>}
   </Typography>
 );
