--- src/components/CustomLabel/label.module.scss
@@ -1,6 +1,0 @@
-@import "../../assets/styles";
-
-.title{
-    font-family: $gotham !important;
-    color: #333333;
-}
