+++ src/layout/Header/Header.module.scss
@@ -19,23 +19,19 @@
 
 .header-text {
   font-family: $gothamBlack !important;
-  font-size: 2.2rem !important;
+  font-size: 1.5rem !important;
   color: $warning;
   font-weight: bolder;
 }
 
 .header-icons {
-  font-size: 1.8rem !important;
+  font-size: 1.2rem !important;
   color: $warning;
-
 }
 
 .header-button {
   background-color: $primary !important;
   color: $tertiary-neutral !important;
-  width: 9rem;
-  height: 2.8rem;
-  font-size: 1.1rem !important;
 }
 
 .header-button-variant {
