+++ src/utils/helpers.ts
@@ -1,5 +1,3 @@
-import dayjs, { Dayjs } from 'dayjs';
-
 export const debounce = <A = unknown, R = void>(
   // source: https://stackblitz.com/github/Bwca/demo__use-debounce-hook?file=src%2Fdebounce%2Fdebounce.ts
   fn: (args: A) => R,
@@ -22,7 +20,3 @@ export const debounce = <A = unknown, R = void>(
 
   return [debouncedFunc, teardown];
 };
-
-export const getFormattedDate = (inputDate: string) => {
-  return dayjs(inputDate).format('MM/DD/YYYY');
-}
