+++ src/views/Profile/components/PersonalInformation2.tsx
@@ -5,16 +5,11 @@ import profileScss from "../../Profile/Profile.module.scss";
 import customTextField from "../sxTexField";
 import { FormikProps } from "formik";
 import { UserProfile } from "@/types/user";
-import { getFormattedDate } from '@/utils/helpers';
-import { CustomTextField } from '../constants';
-import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
-import { DatePicker } from "@mui/x-date-pickers";
 
 interface IPersonalInformation2 {
   isEditMode: boolean;
   formik: FormikProps<UserProfile>
 }
-
 
 const PersonalInformation2 = ({ isEditMode, formik }: IPersonalInformation2) => {
   return (
@@ -24,7 +19,7 @@ const PersonalInformation2 = ({ isEditMode, formik }: IPersonalInformation2) => 
         <>
           <Box sx={{ display: 'flex', paddingBottom: '1rem' }}>
             <Typography>Date or Birth: &nbsp;</Typography>
-            <Typography sx={{ lineBreak: 'anywhere'}}>{getFormattedDate(formik.values.birthdate)}</Typography>
+            <Typography sx={{ lineBreak: 'anywhere'}}>{formik.values.birthdate}</Typography>
           </Box>
           <Box sx={{ display: 'flex', paddingBottom: '1rem' }}>
             <Typography>Phone number: &nbsp;</Typography>
@@ -46,7 +41,7 @@ const PersonalInformation2 = ({ isEditMode, formik }: IPersonalInformation2) => 
               onChange={formik.handleChange}
               error={formik.touched.birthdate && !!formik.errors.birthdate}
               helperText={formik.touched.birthdate && formik.errors.birthdate}
-              sx={{...CustomTextField, width: '50%' }}
+              sx={{...customTextField, width: '50%' }}
               type="date"
             />
           </Box>
@@ -61,7 +56,7 @@ const PersonalInformation2 = ({ isEditMode, formik }: IPersonalInformation2) => 
               helperText={
                 formik.touched.cell_phone && formik.errors.cell_phone
               }
-              sx={{...CustomTextField, width: '50%' }}
+              sx={{...customTextField, width: '50%' }}
             />
           </Box>
           <Box sx={{ display: 'flex', paddingBottom: '1rem', width: '100%' }}>
@@ -73,7 +68,7 @@ const PersonalInformation2 = ({ isEditMode, formik }: IPersonalInformation2) => 
               onChange={formik.handleChange}
               error={formik.touched.email && !!formik.errors.email}
               helperText={formik.touched.email && formik.errors.email}
-              sx={{...CustomTextField, width: '50%' }}
+              sx={{...customTextField, width: '50%' }}
             />
           </Box>
         </>
