+++ src/views/Register/Register.module.scss
@@ -12,18 +12,16 @@
 .submitButton {
   background-color: $primary !important;
   color: $white !important;
-  height: 3rem;
 }
 
 .cancelButton {
   color: $primary !important;
   border-color: $primary !important;
-  height: 3rem;
 }
 
 .title {
   font-family: $gothamBlack !important;
-  font-size: 2.5rem !important;
+  font-size: 2rem;
   color: $warning;
   padding-left: 1rem;
-} 
+}
