--- src/views/Register/components/PersonalInformations/EmailField.tsx
@@ -1,56 +1,0 @@
-import Grid from "@mui/material/Grid";
-import TextField from "@mui/material/TextField";
-import { CustomLabel } from "@/components";
-const EmailField = ({ formik }) => {
-  const primaryColor = "#009999";
-  const placeholderColor = "rgba(51, 51, 51, 0.4)";
-
-  const customTextField = {
-    width: "90%",
-    "& .MuiInputBase-input": {
-      height: "1.9rem",
-    },
-    "& .MuiOutlinedInput-root .MuiOutlinedInput-notchedOutline": {
-      borderColor: primaryColor,
-      borderRadius: 0,
-      border: "2px solid " + primaryColor,
-    },
-    "&:hover .MuiOutlinedInput-root .MuiOutlinedInput-notchedOutline": {
-      borderColor: primaryColor,
-    },
-    "& .MuiOutlinedInput-root.Mui-focused .MuiOutlinedInput-notchedOutline": {
-      borderColor: primaryColor,
-    },
-    "& .MuiInputLabel-outlined": {
-      fontSize: "1rem",
-      color: placeholderColor,
-    },
-    "& .MuiInputLabel-outlined.Mui-focused": {
-      color: primaryColor,
-    },
-    "& .MuiOutlinedInput-input": {
-      padding: "0.7rem",
-    },
-  };
-
-  return (
-    <Grid item xs={12} sm={6} md={4}>
-      <CustomLabel name="Email" required={true} />
-      <TextField
-        sx={customTextField}
-        name="email"
-        type="email"
-        placeholder="Email"
-        onChange={formik.handleChange}
-        value={formik.values.email}
-      />
-         {formik.touched.email && formik.errors.email && (
-        <p style={{ color: "red", fontWeight: "lighter", fontSize:12 }}>
-        {formik.errors.email}
-        </p>
-      )}
-    </Grid>
-  );
-};
-
-export default EmailField;
