--- src/views/Register/components/PersonalInformations/PhoneNumberField.tsx
@@ -1,57 +1,0 @@
-import Grid from "@mui/material/Grid";
-import TextField from "@mui/material/TextField";
-import { CustomLabel } from "@/components";
-const PhoneNumberField = ({ formik }) => {
-  const primaryColor = "#009999";
-  const placeholderColor = "rgba(51, 51, 51, 0.4)";
-
-  const customTextField = {
-    width: "90%",
-    "& .MuiInputBase-input": {
-      height: "1.9rem",
-    },
-    "& .MuiOutlinedInput-root .MuiOutlinedInput-notchedOutline": {
-      borderColor: primaryColor,
-      borderRadius: 0,
-      border: "2px solid " + primaryColor,
-    },
-    "&:hover .MuiOutlinedInput-root .MuiOutlinedInput-notchedOutline": {
-      borderColor: primaryColor,
-    },
-    "& .MuiOutlinedInput-root.Mui-focused .MuiOutlinedInput-notchedOutline": {
-      borderColor: primaryColor,
-    },
-    "& .MuiInputLabel-outlined": {
-      fontSize: "1rem",
-      color: placeholderColor,
-    },
-    "& .MuiInputLabel-outlined.Mui-focused": {
-      color: primaryColor,
-    },
-    "& .MuiOutlinedInput-input": {
-      padding: "0.7rem",
-    },
-  };
-
-  return (
-    <Grid item xs={12} sm={6} md={4}>
-      <CustomLabel name="Phone Number" required={true} />
-      <TextField
-        sx={customTextField}
-        name="phoneNumber"
-        type="text"
-        placeholder="Phone Number"
-        onChange={formik.handleChange}
-        value={formik.values.phoneNumber}
-      
-      />
-          {formik.touched.phoneNumber && formik.errors.phoneNumber && (
-        <p style={{ color: "red", fontWeight: "lighter", fontSize:12 }}>
-        {formik.errors.phoneNumber}
-        </p>
-      )}
-    </Grid>
-  );
-};
-
-export default PhoneNumberField;
